+++ src/Context.php
@@ -26,7 +26,7 @@ class Context implements \ArrayAccess
         public readonly ?float $timeout = null,
         public readonly bool $quiet = false,
         public readonly bool $allowFailure = false,
-        public readonly bool $notify = false,
+        public readonly ?bool $notify = null,
         public readonly VerbosityLevel|LegacyVerbosityLevel $verbosityLevel = VerbosityLevel::NOT_CONFIGURED,
         // Do not use this argument, it is only used internally by the application
         public readonly string $name = '',
@@ -222,7 +222,7 @@ class Context implements \ArrayAccess
         );
     }
 
-    public function withNotify(bool $notify = true): self
+    public function withNotify(?bool $notify = true): self
     {
         return new self(
             $this->data,