+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Import and load task from remote import automatically
 * Fix multiple remote imports of the same package with default version
 * Add `context()` function in expression language to enable a task
+* Fix import of local tasks when using remote import
 
 
 ## 0.15.0 (2024-04-03)