+++ CHANGELOG.md
@@ -5,13 +5,14 @@
 * Add support for SSH connection without specifying a user
 * Marked SSH features as stable
 * Import and load task from remote import automatically
-* Fix multiple remote imports of the same package with default version
+* Allow to use imported class in task from remote import
 * Do not load task from `vendor` directory
 * Add `context()` function in expression language to enable a task
 * Deprecate `Castor\GlobalHelper` class. There are no replacements. Use raw
   functions instead
 * Deprecate `AfterApplicationInitializationEvent` event. Use
   `FunctionsResolvedEvent` instead.
+* Fix multiple remote imports of the same package with default version
 
 ## 0.15.0 (2024-04-03)