+++ CHANGELOG.md
@@ -16,6 +16,10 @@
   `FunctionsResolvedEvent` instead.
 * Add `notificationTitle` property to `Context` to set the application name for
   notifications title
+* Changed the behavior of `notify` parameter in `Context` to be a nullable boolean.
+  - `null` is now the default value (only user notifications are displayed).
+  - `true` to enable notifications globally (user and Castor generated notifications)
+  - `false` to disable them globally
 
 ## 0.15.0 (2024-04-03)